make script useable

+++ config/kubermatic/migrate/migrate-kubermatic-chart.sh
@@ -1,8 +1,21 @@
 #!/usr/bin/env bash
 
-# Modify according to the tiller namespace which contains the release configmaps.
-# Its what gets specified with the "--tiller-namespace" flag on Helm. Default is "kube-system"
-TILLER_NAMESPACE='kube-system'
+if [ "$#" -lt 1 ] || [ "${1}" == "--help" ]; then
+  cat <<EOF
+Usage: $(basename $0) path/to/values.yaml tiller-namespace
+EOF
+  exit 0
+fi
+
+if [[ ! -f ${1} ]]; then
+    echo "File not found!"
+    exit 1
+fi
+
+VALUES_FILE=$(realpath ${1})
+TILLER_NAMESPACE=${2:-'kube-system'}
+cd "$(dirname "$0")/../../"
+
 
 # Delete the kubermatic namespace to enable a clean install afterwards
 kubectl delete ns kubermatic
@@ -15,9 +28,6 @@ for cm in $(kubectl -n kubermatic-installer get configmap -o json | jq -r '.item
     kubectl -n ${TILLER_NAMESPACE} delete configmap ${cm}
 done
 
-helm upgrade --install --kube-context=europe-west3-c --tiller-namespace=kubermatic-installer \
-    --set=kubermatic.controller.image.tag=aee6917710fde8e8b8ca65c73d904f950500af55 \
-    --set=kubermatic.api.image.tag=aee6917710fde8e8b8ca65c73d904f950500af55 \
-    --set=kubermatic.rbac.image.tag=aee6917710fde8e8b8ca65c73d904f950500af55 \
-    --values /home/henrik/go/src/github.com/kubermatic/secrets/seed-clusters/dev.kubermatic.io/values.yaml \
-    --namespace kubermatic kubermatic ./kubermatic/
+helm upgrade --install --tiller-namespace=${TILLER_NAMESPACE} \
+    --values ${VALUES_FILE} \
+    --namespace kubermatic kubermatic ./kubermatic/ --dry-run --debug
